feat(dashboard): Adição de um dashboard simples para o coordenador
