+++ frontend/src/components/Coordenador/Dashboard/DashCoordenador.tsx
@@ -1,6 +1,6 @@
 import { useCoordenador } from "@/contexts/coordenadorContext"
-import { Navbar } from "../Navbar/Navbar"
-import { Card } from "@/components/Dasboard/DashboardCards"
+import { Navbar } from "../../Navbar/Navbar"
+import { Card } from "@/components/Coordenador/Dashboard/DashboardCards"
 
 import { FaChalkboardTeacher, FaGraduationCap } from "react-icons/fa";
 import { IoBookSharp } from "react-icons/io5";
@@ -26,26 +26,21 @@ export const DashCoordenador = () => {
 
 
   return (
-    <div className="flex flex-col h-full gap-15 pt-[1vw] ">
+    <div className="flex flex-col h-full gap-15 max-sm:pt-[12vw] ">
       {/* <Navbar /> */}
-      <div>
-        <div className="flex flex-col gap-10 mb-10">
-          <span className="flex items-center justify-center text-[2vw]">Visão Geral</span>
-          <div className="flex flex-row flex-wrap gap-20 justify-center">
+      
+        <div className="flex h-full flex-col ">
+          <span className="flex items-center justify-center text-[6vw] md:text-[3vw]">Visão Geral</span>
+          <div className="flex flex-col h-full items-center md:flex-row flex-wrap justify-around">
             <Card icon={FaGraduationCap} title={"Alunos"} valueText={overview?.totalAlunos} />
             <Card icon={FaChalkboardTeacher} title={"Professores"} valueText={overview?.totalProfessores ?? null} />
             <Card icon={IoBookSharp} title={"Disciplinas"} valueText={overview.totalDisciplinas ?? null} />
             <Card icon={SiGoogleclassroom} title={"Turmas"} valueText={overview.totalTurmas ?? null} />
           </div>
-          <div className="flex flex-row flex-wrap gap-20 justify-center">
-            <Card icon={FaGraduationCap} title={"Alunos"} valueText={overview?.totalAlunos} />
-            <Card icon={FaChalkboardTeacher} title={"Professores"} valueText={overview?.totalProfessores ?? null} />
-            <Card icon={IoBookSharp} title={"Disciplinas"} valueText={overview.totalDisciplinas ?? null} />
-            <Card icon={SiGoogleclassroom} title={"Turmas"} valueText={overview.totalTurmas ?? null} />
-          </div>
+          
         </div>
 
-      </div>
+      
     </div>
   )
 }
